second commit

+++ icon.h
@@ -13,15 +13,15 @@ public:
     //确定各类物体子素材图片的位置，高、宽等数据
     static ICON findICON(string type);
     //根据物体类型名找到图标
-    ICON(string name,int x,int y,int w,int h);
+    ICON(string name,double x,double y,double w,double h);
     string getTypeName() const{return this->typeName;}
-    int getX() const{return this->X;}
-    int getY() const{return this->Y;}
-    int getW() const{return this->W;}
-    int getH() const{return this->H;}
+    double getX() const{return this->X;}
+    double getY() const{return this->Y;}
+    double getW() const{return this->W;}
+    double getH() const{return this->H;}
 private:
     string typeName;
-    int X,Y,W,H;
+    double X,Y,W,H;
 };
 
 #endif // ICON_H
